# src/Vibetech.Educat.Web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vibetech Educat API",
    "description": "API для платформы репетиторов и студентов",
    "contact": {
      "name": "Vibetech",
      "email": "support@vibetech.ru"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Profile/teacher/{teacherId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Получение профиля преподавателя",
        "description": "Возвращает данные профиля преподавателя по ID",
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль преподавателя",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Преподаватель не найден"
          }
        }
      }
    },
    "/api/Profile/student/{studentId}": {
      "get": {
        "tags": [
          "Profile"
        ],
        "summary": "Получение профиля студента",
        "description": "Возвращает данные профиля студента по ID",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль студента",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Студент не найден"
          }
        }
      }
    },
    "/api/Review/teacher/{teacherId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Получение отзывов репетитора",
        "description": "Возвращает список отзывов, оставленных студентами для указанного репетитора",
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список отзывов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Review/lesson/{lessonId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Получение отзывов об уроке",
        "description": "Возвращает список отзывов, оставленных для указанного урока",
        "parameters": [
          {
            "name": "lessonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список отзывов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Review/student/{studentId}": {
      "get": {
        "tags": [
          "Review"
        ],
        "summary": "Получение отзывов, оставленных студентом",
        "description": "Возвращает список отзывов, оставленных указанным студентом",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список отзывов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Review/create": {
      "post": {
        "tags": [
          "Review"
        ],
        "summary": "Создание отзыва",
        "description": "Создает новый отзыв о репетиторе для указанного урока",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Отзыв успешно создан",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDto"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка при создании отзыва"
          }
        }
      }
    },
    "/api/Student/profile/{studentId}": {
      "get": {
        "tags": [
          "Student"
        ],
        "summary": "Получение профиля студента",
        "description": "Возвращает данные профиля студента по ID",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль студента",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Студент не найден"
          }
        }
      },
      "put": {
        "tags": [
          "Student"
        ],
        "summary": "Обновление профиля студента",
        "description": "Обновляет данные профиля студента",
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профиль студента успешно обновлен",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Студент не найден"
          }
        }
      }
    },
...
    "/api/Teacher/profile/{teacherId}": {
      "get": {
        "tags": [
          "Teacher"
        ],
        "summary": "Получение профиля репетитора",
        "description": "Возвращает данные профиля репетитора по ID",
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Профиль репетитора",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Репетитор не найден"
          }
        }
      },
      "put": {
        "tags": [
          "Teacher"
        ],
        "summary": "Обновление профиля репетитора",
        "description": "Обновляет данные профиля репетитора",
        "parameters": [
          {
            "name": "teacherId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherProfileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherProfileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeacherProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профиль репетитора успешно обновлен",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeacherProfileDto"
                }
              }
            }
          },
          "404": {
            "description": "Репетитор не найден"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateReviewRequest": {
        "required": [
          "comment",
          "lessonId",
          "rating",
          "studentId",
          "teacherId"
        ],
        "type": "object",
        "properties": {
          "lessonId": {
            "type": "integer",
            "format": "int32"
          },
          "teacherId": {
            "type": "integer",
            "format": "int32"
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "maxLength": 2000,
            "minLength": 10,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "PreparationProgramDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "lessonId": {
            "type": "integer",
            "format": "int32"
          },
          "teacherId": {
            "type": "integer",
            "format": "int32"
          },
          "teacherName": {
            "type": "string",
            "nullable": true
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "studentName": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Дата создания отзыва",
            "format": "date",
            "example": "2025-05-28"
          }
        },
        "additionalProperties": false
      },
      "StudentProfileDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "additionalProperties": false
      },
...
      "SubjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TeacherProfileDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "education": {
            "type": "string",
            "nullable": true
          },
          "experienceYears": {
            "type": "integer",
            "format": "int32"
          },
          "hourlyRate": {
            "type": "number",
            "format": "double"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "reviewsCount": {
            "type": "integer",
            "format": "int32"
          },
          "preparationPrograms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PreparationProgramDto"
            },
            "nullable": true
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectDto"
            },
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateStudentProfileRequest": {
        "required": [
          "birthDate",
          "contactInfo",
          "firstName",
          "gender",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "type": "string"
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения в формате YYYY-MM-DD",
            "format": "date",
            "example": "2025-05-28"
          },
          "gender": {
            "minLength": 1,
            "type": "string"
          },
          "contactInfo": {
            "minLength": 1,
            "type": "string"
          },
          "photoBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateTeacherProfileRequest": {
        "required": [
          "birthDate",
          "contactInfo",
          "education",
          "experienceYears",
          "firstName",
          "gender",
          "hourlyRate",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "type": "string"
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения в формате YYYY-MM-DD",
            "format": "date",
            "example": "2025-05-28"
          },
          "gender": {
            "minLength": 1,
            "type": "string"
          },
          "contactInfo": {
            "minLength": 1,
            "type": "string"
          },
          "education": {
            "minLength": 1,
            "type": "string"
          },
          "experienceYears": {
            "maximum": 50,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "hourlyRate": {
            "maximum": 10000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "subjectIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "preparationProgramIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "photoBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения в формате YYYY-MM-DD",
            "format": "date",
            "example": "2025-05-28"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "contactInfo": {
            "type": "string",
            "nullable": true
          },
          "isTeacher": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "photoBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}